Improve names: "CardGroup"==>"CardRow" and "RowOfCards-->CardGroup"
