+++ src/components/CardSwiper.tsx
@@ -11,7 +11,7 @@ import "swiper/swiper.min.css";
 import "swiper/components/navigation/navigation.min.css";
 import "swiper/components/a11y/a11y.min.css";
 import { useResponsiveChoice } from "../responsiveUtilities";
-import { ICardSpec } from "./RowOfCards";
+import { ICardSpec } from "./CardGroup";
 import { commonUI } from "../theme";
 
 SwiperCore.use([Navigation, A11y]);
